+++ Bluejay/Bluejay/Connection.swift
@@ -9,11 +9,6 @@
 import Foundation
 import CoreBluetooth
 
-var standardConnectOptions: [String : AnyObject] = [
-    CBConnectPeripheralOptionNotifyOnDisconnectionKey: true as AnyObject,
-    CBConnectPeripheralOptionNotifyOnConnectionKey: true as AnyObject
-]
-
 /// Types of connection time outs. Can specify a time out in seconds, or no time out.
 public enum Timeout {
     /// Specify a timeout with a duration in seconds.
@@ -39,24 +34,26 @@ class Connection: Queueable {
     
     /// Callback for the connection attempt.
     var callback: ((ConnectionResult) -> Void)?
-    
+
+    /// The warning options to use for this particular connection.
+    let warningOptions: WarningOptions
+
     private var connectionTimer: Timer?
     private let timeout: Timeout?
     
-    init(peripheral: CBPeripheral, manager: CBCentralManager, timeout: Timeout, callback: @escaping (ConnectionResult) -> Void) {
+    init(peripheral: CBPeripheral, manager: CBCentralManager, timeout: Timeout, warningOptions: WarningOptions, callback: @escaping (ConnectionResult) -> Void) {
         self.state = .notStarted
         
         self.peripheral = peripheral
         self.manager = manager
-        
         self.timeout = timeout
-        
+        self.warningOptions = warningOptions
         self.callback = callback
     }
     
     func start() {
         state = .running
-        manager.connect(peripheral, options: standardConnectOptions)
+        manager.connect(peripheral, options: warningOptions.dictionary)
         
         cancelTimer()