+++ Bluejay/Bluejay/Scan.swift
@@ -171,7 +171,7 @@ class Scan: Queueable {
 
                 stopScan(with: discoveries, error: nil)
             }
-            else if case .connect(let discovery, let timeout, let completion) = discovery(newDiscovery, discoveries) {
+            else if case .connect(let discovery, let timeout, let warningOptions, let completion) = discovery(newDiscovery, discoveries) {
                 state = .completed
                 
                 log("Finished scanning.")
@@ -180,7 +180,13 @@ class Scan: Queueable {
                 
                 if let queue = queue {
                     if let cbPeripheral = manager.retrievePeripherals(withIdentifiers: [discovery.peripheralIdentifier.uuid]).first {
-                        queue.add(Connection(peripheral: cbPeripheral, manager: manager, timeout: timeout, callback: completion))
+                        queue.add(Connection(
+                            peripheral: cbPeripheral,
+                            manager: manager,
+                            timeout: timeout,
+                            warningOptions: warningOptions,
+                            callback: completion)
+                        )
                     }
                     else {
                         completion(.failure(BluejayError.unexpectedPeripheral(discovery.peripheralIdentifier)))