+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Added an option to `scan` to change the threshold for ignoring discoveries based on insignificant changes to RSSI
 
 ## [0.6.4] - 2018-08-02
 ### Added